Refactor read only service
