--- lib/appnexusapi/ad_server_service.rb
@@ -1,15 +1,0 @@
-class AppnexusApi::AdServerService < AppnexusApi::Service
-
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-
-  def name
-    "adserver"
-  end
-
-  def uri_suffix
-    "ad-server"
-  end
-end
