--- lib/appnexusapi/device_model_service.rb
@@ -1,6 +1,0 @@
-class AppnexusApi::DeviceModelService < AppnexusApi::Service
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-end
