--- lib/appnexusapi/inventory_attribute_service.rb
@@ -1,8 +1,0 @@
-class AppnexusApi::InventoryAttributeService < AppnexusApi::Service
-
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-
-end
