--- lib/appnexusapi/media_type_service.rb
@@ -1,8 +1,0 @@
-class AppnexusApi::MediaTypeService < AppnexusApi::Service
-
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-
-end
