--- lib/appnexusapi/operating_system_extended_service.rb
@@ -1,10 +1,0 @@
-class AppnexusApi::OperatingSystemExtendedService < AppnexusApi::Service
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-
-  def plural_name
-    'operating-systems-extended'
-  end
-end
