--- lib/appnexusapi/operating_system_service.rb
@@ -1,6 +1,0 @@
-class AppnexusApi::OperatingSystemService < AppnexusApi::Service
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-end
