--- lib/appnexusapi/platform_member_service.rb
@@ -1,8 +1,0 @@
-class AppnexusApi::PlatformMemberService < AppnexusApi::Service
-
-  def initialize(connection)
-    @read_only = true
-    super(connection)
-  end
-
-end
