+++ lib/appnexusapi/read_only_service.rb
@@ -1,0 +1,8 @@
+module AppnexusApi
+  class ReadOnlyService < Service
+    def initialize(connection)
+      @read_only = true
+      super(connection)
+    end
+  end
+end
