+++ lib/appnexusapi/services/ad_server_service.rb
@@ -1,0 +1,9 @@
+class AppnexusApi::AdServerService < AppnexusApi::ReadOnlyService
+  def name
+    "adserver"
+  end
+
+  def uri_suffix
+    "ad-server"
+  end
+end
