+++ lib/appnexusapi/services/bidder_instance_service.rb
