+++ lib/appnexusapi/services/bidder_service.rb
