+++ lib/appnexusapi/services/brand_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::BrandService < AppnexusApi::ReadOnlyService
+end
