+++ lib/appnexusapi/services/browser_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::BrowserService < AppnexusApi::ReadOnlyService
+end
