+++ lib/appnexusapi/services/category_service.rb
@@ -1,0 +1,5 @@
+class AppnexusApi::CategoryService < AppnexusApi::ReadOnlyService
+  def plural_name
+    "categories"
+  end
+end
