+++ lib/appnexusapi/services/content_category_service.rb
