+++ lib/appnexusapi/services/creative_template_service.rb
