+++ lib/appnexusapi/services/device_model_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::DeviceModelService < AppnexusApi::ReadOnlyService
+end
