+++ lib/appnexusapi/services/inventory_attribute_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::InventoryAttributeService < AppnexusApi::ReadOnlyService
+end
