+++ lib/appnexusapi/services/language_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::LanguageService < AppnexusApi::ReadOnlyService
+end
