+++ lib/appnexusapi/services/media_type_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::MediaTypeService < AppnexusApi::ReadOnlyService
+end
