+++ lib/appnexusapi/services/member_service.rb
