+++ lib/appnexusapi/services/operating_system_extended_service.rb
@@ -1,0 +1,5 @@
+class AppnexusApi::OperatingSystemExtendedService < AppnexusApi::ReadOnlyService
+  def plural_name
+    'operating-systems-extended'
+  end
+end
