+++ lib/appnexusapi/services/operating_system_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::OperatingSystemService < AppnexusApi::ReadOnlyService
+end
