+++ lib/appnexusapi/services/payment_rule_service.rb
