+++ lib/appnexusapi/services/placement_service.rb
