+++ lib/appnexusapi/services/platform_member_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::PlatformMemberService < AppnexusApi::ReadOnlyService
+end
