+++ lib/appnexusapi/services/publisher_service.rb
