+++ lib/appnexusapi/services/segment_service.rb
