+++ lib/appnexusapi/services/technical_attribute_service.rb
@@ -1,0 +1,2 @@
+class AppnexusApi::TechnicalAttributeService < AppnexusApi::ReadOnlyService
+end
