+++ lib/appnexusapi/services/user_service.rb
